#250 [rename] 인기 프로젝트, 프로젝트 필터링 어댑터 구현체의 위치를 변경한다.
